Crate vectis_wallet

Modules

factory_state

Structs

Controller

CreateWalletMsg

FeesResponse

Guardians

GuardiansUpdateMsg

If the Guardians.guardian_multisig is given, we will instantiate a new multisig contract. This contract can be an instance of 2 code ids.

GuardiansUpdateRequest

MultiSig

PluginListResponse

PluginParams

ProxyInstantiateMsg

ProxyMigrateMsg

RelayTransaction

WalletCreateReply

WalletFactoryInstantiateMsg

WalletInfo

Enums

CodeIdType

DeployerItemsQueryError

Relay transaction related errors

FactoryError

FeeType

MigrationMsgError

Contract migration related errors

PluginPermissions

Permission of the plugin on the proxy

PluginSource

The source of the plugin code.

ProxyAddrErr

Checks of controller addr

ProxyExecuteMsg

ProxyMigrationTxMsg

ProxyQueryMsg

RelayTxError

Relay transaction related errors

VectisActors

WalletAddr

Representation of the wallet address in both form used in migration

WalletFactoryExecuteMsg

WalletFactoryQueryMsg

Constants

CONTROLLER

Vecits Proxy Controller

DEFAULT_LIMIT

DEPLOYER

In beta-V1 this is a multisig In progressive decentralisation this will be a DAAO

GUARDIAN_REQUEST_ACTIVATION_TIME

Min delay time is set to 1 day

ITEMS

Maps VectisActors variant to addresses: using String as key to keep inline with the dao-core contract

MAX_LIMIT

QUERY_PLUGINS

Vectis Proxy state for other contract to query it

Functions

get_items_from_deployer

pub_key_to_address

Converts controller pubkey into Addr

query_verify_cosmos

Type Definitions

Nonce

Controller nonce

ThresholdAbsoluteCount

Declares that a fixed weight of Yes votes is needed to pass. See ThresholdResponse.AbsoluteCount in the cw3 spec for details. Only Fixed multisig is supported in this version