Trait vectis_wallet::interface::factory_management_trait::FactoryManagementTrait

pub trait FactoryManagementTrait {
    type Error: From<StdError>;

    // Required methods
    fn update_code_id(
        &self,
        ctx: ExecCtx<'_>,
        ty: CodeIdType,
        code_id: u64,
        version: Option<String>,
        set_as_default: bool
    ) -> Result<Response, Self::Error>;
    fn update_config_fee(
        &self,
        ctx: ExecCtx<'_>,
        ty: FeeType,
        new_fee: Coin
    ) -> Result<Response, Self::Error>;
    fn update_supported_interchain(
        &self,
        ctx: ExecCtx<'_>,
        chain_id: String,
        chain_connection: Option<ChainConnection>
    ) -> Result<Response, Self::Error>;
    fn update_deployer(
        &self,
        ctx: ExecCtx<'_>,
        addr: String
    ) -> Result<Response, Self::Error>;
    fn update_wallet_creator(
        &self,
        ctx: ExecCtx<'_>,
        addr: String
    ) -> Result<Response, Self::Error>;
    fn update_auth_provider(
        &self,
        ctx: ExecCtx<'_>,
        ty: AuthenticatorType,
        new_code_id: Option<u64>,
        new_inst_msg: Option<Binary>
    ) -> Result<Response, Self::Error>;
    fn total_created(&self, ctx: QueryCtx<'_>) -> Result<u64, StdError>;
    fn default_proxy_code_id(&self, ctx: QueryCtx<'_>) -> Result<u64, StdError>;
    fn deployer(&self, ctx: QueryCtx<'_>) -> Result<Addr, StdError>;
    fn wallet_creator(&self, ctx: QueryCtx<'_>) -> Result<Addr, StdError>;
    fn supported_chains(
        &self,
        ctx: QueryCtx<'_>,
        start_after: Option<String>,
        limit: Option<u32>
    ) -> Result<Vec<(String, ChainConnection)>, StdError>;
    fn supported_proxies(
        &self,
        ctx: QueryCtx<'_>,
        start_after: Option<u64>,
        limit: Option<u32>
    ) -> Result<Vec<(u64, String)>, StdError>;
    fn fees(&self, ctx: QueryCtx<'_>) -> Result<FeesResponse, StdError>;
    fn auth_provider_addr(
        &self,
        ctx: QueryCtx<'_>,
        ty: AuthenticatorType
    ) -> Result<Option<Addr>, StdError>;
    fn contract_version(
        &self,
        ctx: QueryCtx<'_>
    ) -> Result<ContractVersion, StdError>;

}

The trait for deployer to interact with factory contract

Required Associated Types

type Error: From<StdError>

Required Methods

fn update_code_id( &self, ctx: ExecCtx<'_>, ty: CodeIdType, code_id: u64, version: Option<String>, set_as_default: bool ) -> Result<Response, Self::Error>

Deployer only: update the newest code_id supported by Vectis removes supported ones if version is None; fails if code_id exist and version is Some;

fn update_config_fee( &self, ctx: ExecCtx<'_>, ty: FeeType, new_fee: Coin ) -> Result<Response, Self::Error>

Deployer only: update the fee associated with using Vectis services

fn update_supported_interchain( &self, ctx: ExecCtx<'_>, chain_id: String, chain_connection: Option<ChainConnection> ) -> Result<Response, Self::Error>

Deployer only: update the supported chains

fn update_deployer( &self, ctx: ExecCtx<'_>, addr: String ) -> Result<Response, Self::Error>

Deployer only: update address associated by the deployer role

fn update_wallet_creator( &self, ctx: ExecCtx<'_>, addr: String ) -> Result<Response, Self::Error>

Deployer only: update address associated by the wallet creator role

fn update_auth_provider( &self, ctx: ExecCtx<'_>, ty: AuthenticatorType, new_code_id: Option<u64>, new_inst_msg: Option<Binary> ) -> Result<Response, Self::Error>

Updates the authenticator provider if new_code_id and new_inst_msg is None, the ty assumes to exist and will be removed. Otherwise, it will be added (if does not exist) or migrated

fn total_created(&self, ctx: QueryCtx<'_>) -> Result<u64, StdError>

Returns total wallets created

fn default_proxy_code_id(&self, ctx: QueryCtx<'_>) -> Result<u64, StdError>

Returns existing codeIds of the proxy and others

fn deployer(&self, ctx: QueryCtx<'_>) -> Result<Addr, StdError>

Returns address of the deployer

fn wallet_creator(&self, ctx: QueryCtx<'_>) -> Result<Addr, StdError>

Returns address of the wallet creator

fn supported_chains( &self, ctx: QueryCtx<'_>, start_after: Option<String>, limit: Option<u32> ) -> Result<Vec<(String, ChainConnection)>, StdError>

Returns supported chains

fn supported_proxies( &self, ctx: QueryCtx<'_>, start_after: Option<u64>, limit: Option<u32> ) -> Result<Vec<(u64, String)>, StdError>

Returns supported proxies

fn fees(&self, ctx: QueryCtx<'_>) -> Result<FeesResponse, StdError>

Returns current fee FeeResponse

fn auth_provider_addr( &self, ctx: QueryCtx<'_>, ty: AuthenticatorType ) -> Result<Option<Addr>, StdError>

Returns address of the authenticator

fn contract_version( &self, ctx: QueryCtx<'_> ) -> Result<ContractVersion, StdError>