Trait vectis_wallet::interface::wallet_trait::WalletTrait

pub trait WalletTrait {
    type Error: From<StdError>;

    // Required methods
    fn info(&self, ctx: QueryCtx<'_>) -> Result<WalletInfo, StdError>;
    fn data(
        &self,
        ctx: QueryCtx<'_>,
        key: Binary
    ) -> Result<Option<Binary>, StdError>;
    fn controller_rotation(
        &self,
        ctx: ExecCtx<'_>,
        new_controller: Controller
    ) -> Result<Response, Self::Error>;
    fn auth_exec(
        &self,
        ctx: ExecCtx<'_>,
        transaction: RelayTransaction
    ) -> Result<Response, Self::Error>;
    fn auth_exec_without_plugins(
        &self,
        ctx: ExecCtx<'_>,
        transaction: RelayTransaction
    ) -> Result<Response, Self::Error>;
    fn controller_nonce_update(
        &self,
        ctx: ExecCtx<'_>
    ) -> Result<Response, Self::Error>;
    fn update_data(
        &self,
        ctx: ExecCtx<'_>,
        data: Vec<(Binary, Option<Binary>)>
    ) -> Result<Response, Self::Error>;
}

The trait for each authenticator contract

Required Associated Types

type Error: From<StdError>

Required Methods

fn info(&self, ctx: QueryCtx<'_>) -> Result<WalletInfo, StdError>

Returns the wallet info

fn data( &self, ctx: QueryCtx<'_>, key: Binary ) -> Result<Option<Binary>, StdError>

Returns the data given the key

fn controller_rotation( &self, ctx: ExecCtx<'_>, new_controller: Controller ) -> Result<Response, Self::Error>

Permission: contract self (controller / plugins)

fn auth_exec( &self, ctx: ExecCtx<'_>, transaction: RelayTransaction ) -> Result<Response, Self::Error>

Permission: Open Main exec function and checks for auth from controller

fn auth_exec_without_plugins( &self, ctx: ExecCtx<'_>, transaction: RelayTransaction ) -> Result<Response, Self::Error>

Permission: Open This allows for exec of tx without called any plugins. Allowing for the removal of faulty plugins on the proxy wallet

fn controller_nonce_update( &self, ctx: ExecCtx<'_> ) -> Result<Response, Self::Error>

Permission: factory This is used by the factory in the case the wallet migrates

fn update_data( &self, ctx: ExecCtx<'_>, data: Vec<(Binary, Option<Binary>)> ) -> Result<Response, Self::Error>

Permission: Controller Updates the data stored (auto replace existing)