Module vectis_wallet::types::state

Enums

VectisActors

Constants

AUTHENTICATORS

CONTROLLER

DEPLOYER

ITEMS

Maps VectisActors variant to addresses: using String as key to keep inline with the dao-core contract

PLUGINS

PROXY_DATA

SUBSCRIBERS

SUPPORTEDCHAINS

SUPPORTEDPROXIES

Type Aliases

Authenticators

These are the authenticators types stored on the factory

ControllerState

Deployer

In v1.0.0 this is a multisig In progressive decentralisation this will be a DAAO

Plugins

Types that allow other contracts to query the plugins from the registry

ProxyData

The data blob store in Proxy

Subscribers

Types that allow other contracts to query the subscriber status from the registry

SupportedChains

The supported chains Map<ChainName, ChainConnection>

SupportedProxies

The bytes of the code_id that are supported to the version